Упрощена модель SecUserDto, оптимизирован импорт

+++ runnerlite-rest/src/main/java/ru/runnerlite/services/IUserService.java
@@ -7,5 +7,5 @@ public interface IUserService {
 	
 	ResponseEntity<SecUserDto> register(SecUserDto user);
 	
-	ResponseEntity<SecUserDto> getById(Long id);
+	ResponseEntity<SecUserDto> getById(Integer id);
 }
